pyproject.toml
[project]
name = "project-registry-api"
version = "0.1.0"
description = "FastAPI-based project registry service with PostgreSQL persistence."
authors = [{ name = "Project Registry Team" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy>=2.0.32,<3.0.0",
    "psycopg[binary]>=3.1.19,<4.0.0",
    "alembic>=1.13.2,<2.0.0",
    "python-jose>=3.3.0,<4.0.0",
    "bcrypt>=4.0.1,<5.0.0",
    "pydantic[email]>=2.8.0,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-multipart>=0.0.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2,<9.0.0",
    "httpx>=0.27.0,<1.0.0",
    "ruff>=0.6.2,<0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2,<9.0.0",
    "httpx>=0.27.0,<1.0.0",
    "ruff>=0.6.2,<0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
ignore = ["E203"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.uv]
managed = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "alembic.ini", "README.md", "docker-compose.yml", "Dockerfile"]
